Deprecate support for MariaDB 10.2.6 and older
