+++ src/Platforms/Keywords/MariaDb102Keywords.php
@@ -5,6 +5,8 @@ namespace Doctrine\DBAL\Platforms\Keywords;
 /**
  * MariaDb reserved keywords list.
  *
+ * @deprecated Use {@link MariaDBKeywords} instead.
+ *
  * @link https://mariadb.com/kb/en/the-mariadb-library/reserved-words/
  */
 final class MariaDb102Keywords extends MariaDBKeywords
@@ -12,260 +14,5 @@ final class MariaDb102Keywords extends MariaDBKeywords
     public function getName(): string
     {
         return 'MariaDb102';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getKeywords(): array
-    {
-        return [
-            'ACCESSIBLE',
-            'ADD',
-            'ALL',
-            'ALTER',
-            'ANALYZE',
-            'AND',
-            'AS',
-            'ASC',
-            'ASENSITIVE',
-            'BEFORE',
-            'BETWEEN',
-            'BIGINT',
-            'BINARY',
-            'BLOB',
-            'BOTH',
-            'BY',
-            'CALL',
-            'CASCADE',
-            'CASE',
-            'CHANGE',
-            'CHAR',
-            'CHARACTER',
-            'CHECK',
-            'COLLATE',
-            'COLUMN',
-            'CONDITION',
-            'CONSTRAINT',
-            'CONTINUE',
-            'CONVERT',
-            'CREATE',
-            'CROSS',
-            'CURRENT_DATE',
-            'CURRENT_TIME',
-            'CURRENT_TIMESTAMP',
-            'CURRENT_USER',
-            'CURSOR',
-            'DATABASE',
-            'DATABASES',
-            'DAY_HOUR',
-            'DAY_MICROSECOND',
-            'DAY_MINUTE',
-            'DAY_SECOND',
-            'DEC',
-            'DECIMAL',
-            'DECLARE',
-            'DEFAULT',
-            'DELAYED',
-            'DELETE',
-            'DESC',
-            'DESCRIBE',
-            'DETERMINISTIC',
-            'DISTINCT',
-            'DISTINCTROW',
-            'DIV',
-            'DOUBLE',
-            'DROP',
-            'DUAL',
-            'EACH',
-            'ELSE',
-            'ELSEIF',
-            'ENCLOSED',
-            'ESCAPED',
-            'EXCEPT',
-            'EXISTS',
-            'EXIT',
-            'EXPLAIN',
-            'FALSE',
-            'FETCH',
-            'FLOAT',
-            'FLOAT4',
-            'FLOAT8',
-            'FOR',
-            'FORCE',
-            'FOREIGN',
-            'FROM',
-            'FULLTEXT',
-            'GENERATED',
-            'GET',
-            'GENERAL',
-            'GRANT',
-            'GROUP',
-            'HAVING',
-            'HIGH_PRIORITY',
-            'HOUR_MICROSECOND',
-            'HOUR_MINUTE',
-            'HOUR_SECOND',
-            'IF',
-            'IGNORE',
-            'IGNORE_SERVER_IDS',
-            'IN',
-            'INDEX',
-            'INFILE',
-            'INNER',
-            'INOUT',
-            'INSENSITIVE',
-            'INSERT',
-            'INT',
-            'INT1',
-            'INT2',
-            'INT3',
-            'INT4',
-            'INT8',
-            'INTEGER',
-            'INTERSECT',
-            'INTERVAL',
-            'INTO',
-            'IO_AFTER_GTIDS',
-            'IO_BEFORE_GTIDS',
-            'IS',
-            'ITERATE',
-            'JOIN',
-            'KEY',
-            'KEYS',
-            'KILL',
-            'LEADING',
-            'LEAVE',
-            'LEFT',
-            'LIKE',
-            'LIMIT',
-            'LINEAR',
-            'LINES',
-            'LOAD',
-            'LOCALTIME',
-            'LOCALTIMESTAMP',
-            'LOCK',
-            'LONG',
-            'LONGBLOB',
-            'LONGTEXT',
-            'LOOP',
-            'LOW_PRIORITY',
-            'MASTER_BIND',
-            'MASTER_HEARTBEAT_PERIOD',
-            'MASTER_SSL_VERIFY_SERVER_CERT',
-            'MATCH',
-            'MAXVALUE',
-            'MEDIUMBLOB',
-            'MEDIUMINT',
-            'MEDIUMTEXT',
-            'MIDDLEINT',
-            'MINUTE_MICROSECOND',
-            'MINUTE_SECOND',
-            'MOD',
-            'MODIFIES',
-            'NATURAL',
-            'NO_WRITE_TO_BINLOG',
-            'NOT',
-            'NULL',
-            'NUMERIC',
-            'ON',
-            'OPTIMIZE',
-            'OPTIMIZER_COSTS',
-            'OPTION',
-            'OPTIONALLY',
-            'OR',
-            'ORDER',
-            'OUT',
-            'OUTER',
-            'OUTFILE',
-            'OVER',
-            'PARTITION',
-            'PRECISION',
-            'PRIMARY',
-            'PROCEDURE',
-            'PURGE',
-            'RANGE',
-            'READ',
-            'READ_WRITE',
-            'READS',
-            'REAL',
-            'RECURSIVE',
-            'REFERENCES',
-            'REGEXP',
-            'RELEASE',
-            'RENAME',
-            'REPEAT',
-            'REPLACE',
-            'REQUIRE',
-            'RESIGNAL',
-            'RESTRICT',
-            'RETURN',
-            'RETURNING',
-            'REVOKE',
-            'RIGHT',
-            'RLIKE',
-            'ROWS',
-            'SCHEMA',
-            'SCHEMAS',
-            'SECOND_MICROSECOND',
-            'SELECT',
-            'SENSITIVE',
-            'SEPARATOR',
-            'SET',
-            'SHOW',
-            'SIGNAL',
-            'SLOW',
-            'SMALLINT',
-            'SPATIAL',
-            'SPECIFIC',
-            'SQL',
-            'SQL_BIG_RESULT',
-            'SQL_CALC_FOUND_ROWS',
-            'SQL_SMALL_RESULT',
-            'SQLEXCEPTION',
-            'SQLSTATE',
-            'SQLWARNING',
-            'SSL',
-            'STARTING',
-            'STORED',
-            'STRAIGHT_JOIN',
-            'TABLE',
-            'TERMINATED',
-            'THEN',
-            'TINYBLOB',
-            'TINYINT',
-            'TINYTEXT',
-            'TO',
-            'TRAILING',
-            'TRIGGER',
-            'TRUE',
-            'UNDO',
-            'UNION',
-            'UNIQUE',
-            'UNLOCK',
-            'UNSIGNED',
-            'UPDATE',
-            'USAGE',
-            'USE',
-            'USING',
-            'UTC_DATE',
-            'UTC_TIME',
-            'UTC_TIMESTAMP',
-            'VALUES',
-            'VARBINARY',
-            'VARCHAR',
-            'VARCHARACTER',
-            'VARYING',
-            'VIRTUAL',
-            'WHEN',
-            'WHERE',
-            'WHILE',
-            'WINDOW',
-            'WITH',
-            'WRITE',
-            'XOR',
-            'YEAR_MONTH',
-            'ZEROFILL',
-        ];
     }
 }
